remove fmtr status from init(), include container runtime info in version cmd
