+++ cmd/hof/verinfo/verinfo.go
@@ -14,14 +14,16 @@ var (
 	BuildOS   = "Unknown"
 	BuildArch = "Unknown"
 
-	CueVersion = "0.5.0"
+	// todo, look this up from deps
+	CueVersion = "0.6.0-beta.1"
 
 	// this is a version we can fetch with hof mod
 	// the value gets injected into templates in various places
 	// the default here is set to something useful for dev
 	// the release version is the same as the cli running it
-	HofVersion = "v0.6.8-beta.10"
+	HofVersion = "v0.6.8-rc.5"
 )
+
 
 func init() {
 	info, _ := debug.ReadBuildInfo()
